requirements.txt
alpaca-py==0.28.0
annotated-types==0.7.0
blinker==1.6.3
cachetools==5.3.3
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
distlib==0.3.7
filelock==3.12.4
Flask==3.0.0
fonttools==4.53.1
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.137.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.63.2
grpcio==1.64.1
grpcio-status==1.62.2
httplib2==0.22.0
idna==3.6
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.9.1
msgpack==1.0.8
numpy==2.0.0
packaging==24.1
pandas==2.2.2
pillow==10.4.0
platformdirs==3.11.0
proto-plus==1.24.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.8.2
pydantic_core==2.20.1
pygame==2.5.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
rsa==4.9
setuptools==68.2.2
six==1.16.0
sseclient-py==1.8.0
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.1.0
virtualenv==20.24.5
websockets==12.0
Werkzeug==3.0.1
